Put more functionality in Connector class <organisation
